Fixes background image gradient for large sreens.
Previously, the fix helped on small screens that needed to scroll. That ended up introducing a bug for large screens which which the page content didn't take the entire page. In that case, the background image would go beyond the contet, but the gradient would not cover the whole thing.

Making the `html` `height: 100%` seems to fix this issue.