Creates sub-directory with Gastby site.
Generates a Gatsby site using [this
template](https://www.gatsbyjs.org/starters/mhadaily/gatsby-starter-typescript-power-blog/)
to try out as a potential replacement for Jekyll.

This also include `.tool-versions` updates because Gatsby is a JS
framework and need node/npm things, and some npm package the started
project included was built unsing python and seemed to die with a v3+,
so we needed to include a 2.7 version.