Changes blog post rendering.
- Updates blog post directory structure
  - Uses a year, month, title directory structure
    - The post itself can be title `index.md`
    - Since each post gets its own directory, we can now have the images
for that post in the same directory, instead of a shared public
directory
  - Removes the sample blog posts
  - Copies over the most recent blog post from existing blog
- Updates blog post front matter
  - Makes the `categories` an array that Gatsby's graphql endpoint will
give as an actual array
  - For existing post, updates the timestamp to be what Gastby (?) wants
for parsing
- Styles blog posts
  - Since Remark parses the markdown and generates the HTML, it seemed
like the easiest way to style it would be to use a CSS file. I made a
`post.css` file and prefixed everything with `.blog-post-body` to limit
the scope of impact. This file used `@apply` to apply Tailwind classes
  - A lot of style was to add spacing around block elements
  - Tailwind 1.0 removed styling for links and lists; for the post
having "normal" styling for these types of elements made sense, so
`post.css` styles them in a pretty traditional way
  - Styles various size headers
  - Brings in [Prism](https://prismjs.com/) for code styling using [the
Gatsby plugin](https://www.gatsbyjs.org/packages/gatsby-remark-prismjs/?=prisma)

There are definitely going to be more style changes that need to be made
(e.g., the bottom section of the blog post that let's you move between
posts) as more posts are brought in. This seemed like a good place to
commit since a lot happened.