Updates `post.css` to be more restrictive to the post body.
The previous rules didn't have `blog-post-body` prefixing all of the top
rules. This was causing the `Blog` list page to have more bottom margin
than intended.

This changes the `post.css` rules to prefix to be scoped in the
`.blog-post-body`.

In addition, adds margin between the post's title/h1, and the body.