Copies more blog posts to Gastby version.
Copies additional blog posts to Gastby version. When copying a blog post
with gifs, I found that `gatsby-remark-images` does not coy over gifs
(or svgs) because the cannot (or do not need to be) optimized:

https://github.com/gatsbyjs/gatsby/issues/7317

As per the suggestion, this commit also adds
`gatsby-remark-copy-linked-files` which will copy over gifs and render
them.

Also, this commit slightly edits the "Using Vim to Drive TDD" post to
remove a link that was broken (using an old HipChat link).