Sets up skeleton of Talks page.
- Creates a Talk page component
- Creates components for the year list and talk itself
- Does very basic styling
- Creates a type declaration file since the Talk and TalkList types are
used in a few different files

Currently, this has been built out using some test data that is created
as a constant in the talk page component. I'd like to try something like
[this](https://www.gatsbyjs.org/docs/gatsby-config/#mapping-node-types)
where you read in JSON from a file and  can query with graphql. That
should allow to avoid the `sort` and `reverse` work.