Updates how Talks page gets its data.
Updates the Talks page to source data from the file system via
`gatsby-transformer-json`.

Using [this
doc](https://www.gatsbyjs.org/docs/sourcing-from-the-filesystem/) as
inspiration and the JSON transformer docs, this commit updates how the
Talks page gets its data.

- Creates a `data` directory in content
- Updates `gatsby-config` to source the `data` directory
- Adds the JSON transformer plugin to tranform the new data JSON file

The JSON transformer assumed the JSON was an array, so I had to change
the structure of the data a bit - previously we were expecting an object
with the key of year and the value was an array of talks. Now, we have
an array of objects with `year` and talks` keys. This resulted in
deleting the `TalkList` type, and replacing it with the `TalkPageProps`
and `TalkNode` types which map to the response from GraphQL.

Because we are able to query from GraphQL, we can sort/order by year
which is nice.