Updates blog post HTML to work with Safari Reader.
Using [this post](https://www.client9.com/designing-html-for-safari-reader-mode/) updates the template for blog posts to work better with HTML.

The main change that seemed to help was the additon of the `main` tag (`article` didn't work, but. according to the post, `main` is more accessible anyway).

This change also include wrapping the posted date in a `time` node. I'm hoping the string format for the `dateTime` attirbute works okay and that this is more acessible for screen readers.