Tests alternative version syntax for GitHub action.
The GitHub action to publish Gatsby has [an issue with version names in
releases](https://github.com/enriikke/gatsby-gh-pages-action/issues/9)
that is resulting in `@v2` to use an old version. This old version does
not correctly deploy when using NPM, so we need to use a newer version.

This commit tries to remove the `v` to see if it will use `@2` to find
releases with `2.X` (no V). If this does not work, we can try the
`@branch` syntax.