Sets Gatsby publish action to use `master` branch.
In a97c53e we tried to workaround an [issue with version
numbers](https://github.com/enriikke/gatsby-gh-pages-action/issues/9) in
the Gatsby publish action, but it did not work. Using the `@branch`
syntax should help us resolve the issue, but could result in pain down
the road since we will be pulling from `master` instead of a set
version.