(fix) Fix issue where static queries weren't running
After upgrading the TS gatsby plugin, when building the site static
queries didn't seem to run and would show up in the HTML as `Loading
(StaticQuery)`.

[This issue](https://github.com/gatsbyjs/gatsby/issues/26038) seemed to
have a similar issue and it seemed _maybe_ related to one of the
following:

* Babel cached files
* Browser caching
* Dependency resolution
* ??

When upgrading Gatsby, and testing the `build` locally, it seems to work
now.