Update header to always fetch SiteName
Originally, I didn't have the SiteName in the headers on the home page.
Due to the size of the site name, this would cause the other elements to
shift when switching between the home page and other pages on the site.

Instead, we will now always include the site name to keep the header
standard.

Since we are now keeping the site name, we could move the `StaticQuery`
to fetch it from the Gatsby site metadata out of the `Layout` and intot
the header iteself.